examples/pyproject.toml
[tool.poetry]
name = "examples"
version = "0.1.0"
description = ""
authors = ["Olli-Pekka Kahilakoski <olli-pekka.kahilakoski@aalto.fi>"]

[tool.poetry.dependencies]
python = "^3.7"
nest-asyncio = "^1.5.1"
python-socketio = {extras = ["client"], version = "^5.4.0"}
requests = "^2.26.0"
uvicorn = {extras = ["standard"], version = "^0.15.0"}
aioconsole = "^0.3.2"
python-dotenv = "^0.19.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
